Done for the night!
#5u8wba
#5u8wbb
#5rd11t

+++ CyberCore/Manager/Factions/FactionLocalCache.cs
@@ -41,7 +41,7 @@ namespace CyberCore.Manager.Factions
 
         }
 
-        public int getTimeToInt(){
+        public long getTimeToInt(){
             return CyberUtils.getLongTime();
         }
     }
